Merge 'Prefault memory when --lock-memory 1 is specified' from Avi Kivity
To avoid latency spikes due to page faults on anonymous memory,
we mlockall() all memory. This prevents memory from being swapped out,
but we can still see a latency spike when faulting in the memory the first time
it is touched.

It's rare for this to be a problem, as faulting in an anonymous memory page
on first use is cheap - the kernel just has to zero it. However, things are complicated
by transparent hugepages as the kernel first has to defragment memory. This can
take quite a while, as was observed in [1].

The solution is to launch background threads that will attempt to fault-in the
memory ahead of the application. We need to be careful so that these threads
themselves don't compete with the application and cause latency spikes themselves,
so we take the following steps:
1. We launch just one thread per NUMA node, reducing lock contention
2. We place the threads in the SCHED_IDLE class, so the kernel will favor application threads
3. We let the thread affinity float over the entire NUMA node, so it can find the least contended core

Tested on an old Intel E5 E5-2697. It was able to fault 200GB in 50 cpu seconds
(25 wall-clock seconds) for a rate of 8 GB/s. This shows that even a large-memory application
will be able to fault in memory faster than it will need it.

[1] https://github.com/scylladb/scylladb/issues/8828

\Closes scylladb/seastar#1702

* https://github.com/scylladb/seastar:
  smp: wire up memory prefaulter
  smp: introduce memory prefaulter
  resource: compute numa_node_id to cpuset mapping in resources::allocate()
  posix: add posix_thread attribute for thread affinity
  memory: return NUMA layout of allocated memory during initialization

(cherry picked from commit 2b7a341210cfd18650c10687e7cf539e804e2df4)