rpc: let compressors send something over the connection
Fancy compressors might want to sends some information to the peer
(e.g. to offer a new dictionary) in some finite time.
But currently, there is no interface (or any mechanism) to make this
possible. (In particular, it's normal for a connection to be entirely
one-sided, giving the receiver no opportunity to communicate anything
to the sender).

This patch adds such a mechanism. For every connection, a condition_variable
is created, and a coroutine is spawned. The condition_variable is passed
to the compressor, and the coroutine sends an empty frame
over the connection when the condition_variable is signalled.
Thus, if a compressor wants to send something to the peer, it can signal
the variable and then add its metadata to the resulting message.