rpc: fix a deadlock in connection::send()
If send_entry() throws (e.g. because compression is unable to allocate
an output buffer), then the intended error handling never gets to run.
In particular, the connection isn't aborted, and the promise of outgoing_entry
is never completed -- the connection is stuck.

The fix is to extend the exception catching back to the send_entry() call.