rpc: fix a deadlock in connection::send()
If send_entry() throws (e.g. because compression is unable to allocate
an output buffer), then the intended error handling never gets to run.
In particular, the connection isn't aborted, and the promise of outgoing_entry
is never completed -- the connection is stuck.

To fix this, we make send_entry() noexcept.